tests: Add missing fake servers
The fake service clients rely on the fake server being initialised, even
if we don't actually use it. Add the missing calls, renaming some tests
in the process.

PS: You can reproduce this without the '-shuffle on' by simply deleting
any of the test cases preceding those that we are modifying here and
running the tests.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>